pyproject.toml
[tool.poetry]
name = "iridium"
version = "0.1.0"
description = "Enhanced wrapper API around InvenioRDM record API."
authors = ["a.pirogov <a.pirogov@fz-juelich.de>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
typer = "^0.4.0"
pydantic = "~1.8.2"
httpx = "^0.21.1"
python-dotenv = "^0.19.2"
typing-extensions = "^4.0.1"
jsonschema = "^4.4.0"
"ruamel.yaml" = "^0.17.20"

[tool.poetry.dev-dependencies]
pre-commit = "^2.13.0"
pytest = "^6.2.4"
pdoc = "^8.0.1"
pytest-cov = "^3.0.0"
pytest-dotenv = "^0.5.2"

[tool.poetry.scripts]
iridium-cli = "iridium.cli:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# pytest.ini replacement (pytest >= 6.0)
[tool.pytest.ini_options]
addopts = "--cov-report=term-missing:skip-covered"
filterwarnings = ["ignore::DeprecationWarning:somepackage.*"]

# .coveragerc replacement
[tool.coverage.run]
source = ['iridium']
